# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TeamUp API
  description: |
    Sports venue booking and team formation API

    Features:
    - Venue search with geolocation
    - Timeslot booking
    - TeamUp creation and management
    - Join request workflow
  version: 1.0.0
  contact:
    name: TeamUp API Support
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.teamup.example.com
    description: Production server
...
paths:
...
  /venues:
    get:
      tags: [Venues]
      summary: Search venues with available timeslots
      description: |
        Search venues based on location, datetime, and sport type.
        Supports geolocation-based search with distance filtering.
      operationId: searchVenues
      parameters:
        - name: lat
          in: query
          description: Latitude coordinate
          schema:
            type: number
            format: float
            example: 25.0330
        - name: lng
          in: query
          description: Longitude coordinate
          schema:
            type: number
            format: float
            example: 121.5654
        - name: distance
          in: query
          description: Search radius in meters (default 5000m)
          schema:
            type: number
            format: float
            default: 5000
            example: 5000
        - name: datetime
          in: query
          description: ISO datetime (YYYY-MM-DDTHH:MM:SS) or date (YYYY-MM-DD)
          schema:
            type: string
            example: "2025-10-20"
        - name: sport_type
          in: query
          description: Filter by sport type
          schema:
            type: string
            example: basketball
      responses:
        '200':
          description: List of venues with available timeslots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VenueSearchResult'
  /venues/{venue_id}:
    get:
      tags: [Venues]
      summary: Get venue details
      operationId: getVenueById
      parameters:
        - name: venue_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Venue details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VenueDetail'
        '404':
          description: Venue not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login/signup
  schemas:
...
    VenueSearchResult:
      type: object
      required: [venue, time_slots]
      properties:
        venue:
          $ref: '#/components/schemas/VenueOut'
        time_slots:
          type: array
          items:
            $ref: '#/components/schemas/TimeSlotOut'
    VenueOut:
      type: object
      required: [id, name, address]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        city:
          type: string
          nullable: true
        latitude:
          type: number
          format: float
          nullable: true
        longitude:
          type: number
          format: float
          nullable: true
        distance_meters:
          type: number
          format: float
          nullable: true
          description: Distance from search coordinates (only when lat/lng provided)
    VenueDetail:
      type: object
      required: [id, name, address, courts, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        city:
          type: string
          nullable: true
        latitude:
          type: number
          format: float
          nullable: true
        longitude:
          type: number
          format: float
          nullable: true
        contact_phone:
          type: string
          nullable: true
        partner_code:
          type: string
          nullable: true
        courts:
          type: array
          items:
            $ref: '#/components/schemas/CourtOut'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CourtOut:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        sport_type:
          type: string
          nullable: true
    TimeSlotOut:
      type: object
      required: [id, court_id, starts_at, ends_at, currency, is_bookable]
      properties:
        id:
          type: string
          format: uuid
        court_id:
          type: string
          format: uuid
        court_name:
          type: string
          nullable: true
          description: Only present in search results
        sport_type:
          type: string
          nullable: true
          description: Only present in search results
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        price_cents:
          type: integer
          nullable: true
        currency:
          type: string
          example: TWD
        is_bookable:
          type: boolean

    # ========================================================================
    # BOOKING SCHEMAS
    # ========================================================================
...